Move OfflineAudioContext to separate file
